Rename wallet package to transaction package
