+++ src/main/java/seedu/address/model/transaction/Amount.java
@@ -1,4 +1,4 @@
-package seedu.address.model.wallet;
+package seedu.address.model.transaction;
 
 /**
  * Represents an Amount of money.
