+++ src/main/java/seedu/address/model/transaction/Income.java
@@ -1,4 +1,4 @@
-package seedu.address.model.wallet;
+package seedu.address.model.transaction;
 
 import java.time.LocalDate;
